Expose device subkey kid in currentSession API call
Some clients need this. (They'll also eventually need the private key,
but that can come later.)

Add separate AlgoType and KIDMapKey types to make it easier to verify
correctness (as opposed to plain strings/ints).

Remove GetActiveSibkeyKidForCurrentDevice() function in favor of
GetSibkeyForDevice().

Since nothing was using GetDeviceKID(), repurpose it and rename it to
GetDeviceSubkeyKid().

This closes #405.