get rid of FOKIDs!
The "fingerprint or key ID" abstraction was invented back when PGP keys
didn't have KIDs. Now that they do, it causes way more problems than it
solves. Lots of code implicitly assumes that one or the other will be
present in a FOKID (especially checking for equality), which isn't
enforced at all by the type, so things tend to break in confusing ways.

With this diff, we now use KIDs everywhere. In the places where
fingerprints are still needed, they're either passed explicitly or
retrieved from the KeyFamily's kid2pgp map. I also went ahead and
renamed "GetKid" to "GetKID" in several places, in keeping with Go's
convention for acronyms.

Closes https://github.com/keybase/client/issues/418.