kex2: RPC layer
This (squashed) commit is the new RPC layer for KEX2, with some pretty good tests.
It relies on #1015 (the KEX2 transport) and also a new version of maxtaco/go-framed-msgpack-rpc,
which we're soon to upgrade to keybase/go-framed-msgpack-rpc

It might still need some more tests, but I wonder if we should continue to build on top
of this library before we add more.

Closes #1050