replace `keybase reset` with `keybase deprovision`
https://keybase.atlassian.net/browse/CORE-319

Prior to this there were a number of different commands to "delete all
my local state", and none of them was really complete. User info would
stick around in config files, and also there was nothing to delete
secret keys. The deprovision command now does all this, and because it's
super unsafe, it gives the user a big scary warning before it runs.