CLI: keybase saltpack encrypt now uses identify2
- make an Identify2+Resolve all-in-one engine
- rename identify2 test now that we have ResolveThenIdentify2 also covered
- identify2 + salt pack encryption integration

- Temporary fix for a bug: which is that if we issue something like:

   keybase exp encrypt t_tracy@rooter+keybase.io@web

We'll get an eager failure, the client will exit, but the service will still
make additional RPCs on a dead channel.

The problem is that due to bugs in the RPC library, this will hang indefinitely,
and the service will be deadlocked since it holds a singleflight lock for the
user as it's waiting for an RPC reply that will never come.