pass the notify_team query param when creating a git repo
This used to be implicit, but it's easier to respect the "Notify the
team" checkbox if we make it explicit. We can do this, because KBFS is
dropping the feature where repos can magically be created just by
pushing them (which would not have gone through our create repo RPC).