Make invites obsolete during teamchain playback (#10171)
During team chain replaying, when we encounter ChangeMembership link,
we are going to take all UIDs from that link, and move all keybase-
type ActiveInvites for these UIDs to ObsoleteInvites. This way, if
there was an invite for a user and later this user was added using
ChangeMembership link, this invite will no longer by seen in e.g. SBS
handler.