change the EK protocol, sign over existing keys
We're going to need to be able to verify teamEKs other than the most
recent one, when we start doing pairwise MAC'ing in chat. That means
the latest PTK needs to sign over all still-valid generations of the
teamEK, not just the most recent. So instead of this sig:

{
  team_ephemeral_dh_public: ...
  generation: ...
  hash_meta: ...
}

We now need this one:

{
  current_team_ek: {
    team_ephemeral_dh_public: ...
    generation: ...
    hash_meta: ...
    ctime: ...  # adding this in to save the client a fetch
  },
  existing_team_eks: [
    { ... older teamEK blob like the one above ... },
    ...
  ]
}

Make this change, and a similar change for userEKs and deviceEKs, for
consistency.

This depends on https://github.com/keybase/keybase/pull/2235, and tests
will be broken until both are landed.