Prime TeamName -> TeamID cache resolutions as a side effect of fetching favorites (#14193)
* cache TeamName -> TeamID resolutions as a side effect of fetching favorites

- when a user does an `ls` in `/keybase/team`, it maps to a GetFavorites call,
and then, in common cases, an immediate call tio Resolve3 on every team in the
list. With this PR, we plumb the team ID back from the server (it was already being
sent down) and then prime the resolver cache with these resolutions. It should hopefully
immediately return in this case, rather than incuring a server round trip.

* actually hit the cache

* fix flow

* one more fix

* Revert "one more fix"

This reverts commit 5729ea8b1297d105cfd4d2c14c750b6e6b2eb5b8.

* Revert "fix flow"

This reverts commit 2e32aefee341075d0f7790426e5fcd0d342c9019.

* make team_id optional

- this means we shouldn't have to change flow at all