Contact deduplication during search and tests (#18596)
Change how contact filtering is done. We used to filter and deduplicate
contacts during syncing, but we are going to move that to "recommendations" 
and "search" RPCs instead.

This fixes some components not being findable in search because they've been
deduplicated away. While we want to remove all but one contacts if they resolve
to the same Keybase user to present in recommendations list, we don't want them
completely gone - user should still be able to find phone numbers or e-mails of
the components in search.

Add more tests for contact syncing and searching with the use of new mocks.