Show service summaries in recommendations in team-building (#20236)
Move ServiceMapper to G so it's available in InterestingPeople RPC
handler.

Because we want this RPC not to make any extra network calls, we use
ServiceMapper with "no network budget", so it will only use what's in cache.
However this is a problem, because unless you've had a user in contacts, you
won't have their service summary in cache. This is solved by caching service
summaries coming in libkb/loaduser.go (in idTable.remoteProofLinks). Because
InterestingPeople returns users you either chatted with or follow, you usually
had to interact with them beforehand, and therefore load them.