Bulk search emails and phone numbers (#23224)
Add new RPC `bulkEmailOrPhoneSearch` to search a comma or newline separated
list of e-mails or a list of formatted phone numbers (service does not have an
ability to validate and format phone numbers, so it expects GUI to do so).

Refactor usersearch.go, add `searchEmailsOrPhoneNumbers` function that does a
batch search for a list of phone numbers and list of e-mails, returning
formatted assertions and search results. This is what `imptofuSearch` used to
do but for one phone number or one e-mail. It now simply calls
`searchEmailsOrPhoneNumbers`.